# docs-new/api/openapi.yaml
openapi: 3.0.3
info:
  title: Limuru Cottage Hospital Queue Management API
  description: |
    REST API for the Limuru Cottage Hospital Queue Management System.
    
    This API provides endpoints for managing patient queues, appointments,
    clinical documentation, and staff communication.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Public endpoints: 30 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Auth endpoints: 5 attempts per 15 minutes
    
    ## Localization
    The API supports English (en) and Swahili (sw) via the Accept-Language header.
  version: 2.0.0
  contact:
    name: Limuru Cottage Hospital IT
    email: it-support@limuruhospital.co.ke
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-logo:
    url: https://limuruhospital.co.ke/logo.png
servers:
  - url: https://api.limuruhospital.co.ke
    description: Production server
  - url: https://api-staging.limuruhospital.co.ke
    description: Staging server
  - url: http://localhost:8787
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/departments:
    get:
      tags: [Departments]
      summary: List departments
      description: Get all active departments
      operationId: listDepartments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Department list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentListResponse'
    post:
      tags: [Departments]
      summary: Create department
      description: Create a new department
      operationId: createDepartment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
  /api/departments/{departmentId}:
    get:
      tags: [Departments]
      summary: Get department
      description: Get department details
      operationId: getDepartment
      security:
        - bearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
    put:
      tags: [Departments]
      summary: Update department
      description: Update department details
      operationId: updateDepartment
      security:
        - bearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
      responses:
        '200':
          description: Department updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'

  # ==================== DOCTORS ====================
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users
      description: Get list of staff users
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, doctor, nurse, receptionist, pharmacist, lab_tech, facility_manager, it_support]
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'

    post:
      tags: [Admin]
      summary: Create user
      description: Create a new staff user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/admin/users/{userId}:
    put:
      tags: [Admin]
      summary: Update user
      description: Update user details
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    delete:
      tags: [Admin]
      summary: Deactivate user
      description: Deactivate a user account
      operationId: deactivateUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/admin/settings:
    get:
      tags: [Admin]
      summary: Get system settings
      description: Get system configuration
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingsResponse'

    put:
      tags: [Admin]
      summary: Update settings
      description: Update system settings
      operationId: updateSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingsResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, doctor, nurse, receptionist, pharmacist, lab_tech, facility_manager, it_support]
        department:
          type: string
        room:
          type: string

    # ============ Queue ============
...
    CreateDepartmentRequest:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
          maxLength: 100
        code:
          type: string
          maxLength: 20
          pattern: '^[A-Z]{3,10}$'
        description:
          type: string
        color:
          type: string
          example: "#FF5733"
        icon:
          type: string
    UpdateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
    DepartmentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Department'
    Department:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        color:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DepartmentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Department'

    # ============ Doctors ============
...
    CreateUserRequest:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, doctor, nurse, receptionist, pharmacist, lab_tech, facility_manager, it_support]
        department:
          type: string
        room:
          type: string
        phone:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
        department:
          type: string
        room:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
    SystemSettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            clinicName:
              type: string
            clinicAddress:
              type: string
            waitTimePerPatient:
              type: integer
              description: Minutes
            defaultPassword:
              type: string
    UpdateSettingsRequest:
      type: object
      properties:
        clinicName:
          type: string
        clinicAddress:
          type: string
        waitTimePerPatient:
          type: integer
        defaultPassword:
          type: string

    # ============ TV Display ============
...